pyproject.toml
[project]
name = "bitfit-nlp-cub"
version = "0.1.0"
description = "PEFT comparison on GLUE tasks: BitFit, LoRA, Prompt-tuning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "transformers>=4.35",
    "datasets",
    "evaluate",
    "peft",
    "hydra-core",
    "omegaconf",
    "wandb",
    "mlflow",
    "scikit-learn",
    "pandas",
    "matplotlib",
    "accelerate",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
